feat(tests): Refactor test project structure + Add test vault
- Added new files: app.json, appearance.json, core-plugins-migration.json, core-plugins.json, workspace.json, README.md, Hidden Note.md
- Renamed ObsidianTextTests.cs to Core/ObsidianTextTests.cs
- Updated Nodsoft.MoltenObsidian.Tests.csproj to include a reference to Nodsoft.MoltenObsidian.Vaults.FileSystem.csproj and added Assets/** to the content list in the project file
- Modified ObsidianText.cs documentation
