+++ Nodsoft.MoltenObsidian.Tests/Assets/TestVault/.obsidian/appearance.json
@@ -1,0 +1,3 @@
+{
+  "accentColor": ""
+}
