+++ Nodsoft.MoltenObsidian.Tests/Assets/TestVault/README.md
@@ -1,0 +1,7 @@
+# README
+
+This is a test vault, meant to be used with the `Nodsoft.MoltenObsidian.Tests` project's test suite.
+
+## See also :
+
+Check out this #cool_tag, and my [[Hidden Note]].
