+++ Nodsoft.MoltenObsidian.Tests/Assets/TestVault/VeryNiceFolder/Hidden Note.md
@@ -1,0 +1,3 @@
+This is a hidden note!
+Not at all.
+It's just in a folder.
